+++ changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-Last change: 26-MAY-2026 CEST
+Last change: 27-MAY-2026 CEST
 
 ## 0.10.1
 
@@ -21,6 +21,7 @@ This patch adds a number of new, useful features rather than simply fixing the b
 - Efficient Filter: much lower RAM usage + faster filtering by histogram selection (1D/2D), especially for very large datasets
 - Fixed Gauss-fitting error when spot's sum is zero (zero division error)
 - Fixed NND plot reindexing after fitting (#665)
+- Faster cluster centers calculation for SMLM clusterer, DBSCAN and HDBSCAN + `lpz` is saved if applicable
 
 ## 0.10.0